+++ qemu/target-i386/translate.c
@@ -8948,7 +8948,7 @@ void gen_intermediate_code(CPUX86State *env, TranslationBlock *tb)
     TCGContext *tcg_ctx = env->uc->tcg_ctx;
     DisasContext dc1, *dc = &dc1;
     target_ulong pc_ptr;
-    uint64_t flags;
+    uint32_t flags;
     target_ulong pc_start;
     target_ulong cs_base;
     int num_insns = 0;
